backend/requirements.txt
flask
torch
torchvision
pandas
scikit-learn
joblib
Pillow
